feat(cli): Initialize CLI project
This commit initializes the CLI project, adding it to the solution file,
and adding `Tsk.Domain` and `Tsk.Infrastruccture` as references.

The CLI will need to access the user's default application storage path.
This commit adds an `AppData` with the `GetDefaultTskPath()` static
method that returns this path based on the user's OS.