feat(infra): Add TodoRepositoryFactory (#14)
This commit adds the `ITodoRepositoryFactory` interface, as well as its
implementation, the `FlatFileTodoRepositoryFactory`.

The need for this became apparent once I explored `Spectre.Console.Cli`.
A feature of Tsk is for the user to optionally specify a custom file
location. This means that the TodoRepository cannot be initialized
before Spectre.Console has had a chance to parse the command line
arguments.

Thankfully, there is a workaround, and that is to use a DI Container,
and to pass in a factory for the repository, so that the repository can
be instantiated with the correct path.