feat(tests): Add tests for `Tsk.CLI` commands (#27)
* refactor(cli): Swap manual `DateOnly` logic with
`Parser.ParseDueDateFromString`

The `ListCommand` had this helper/parser method as well; this brings the
two in alignment.

* feat(tests): Add tests for `Tsk.CLI` commands

In the course of testing, it was discovered that it was impossible to
delete metadata (e.g. location, due date). Thus, this commit also adds
the option to pass blank strings (`""`) as arguments to the options to
remove those metadata items.